+++ internal/decoders/branch.go
@@ -55,7 +55,7 @@ func formatBranchTarget(target uint32) string {
 func decodeJSR(data []byte, opcode uint16, inst *Instruction) error {
 	mode := uint8((opcode >> 3) & 0x7)
 	reg := uint8(opcode & 0x7)
-	operand, offset, meta, err := decodeEA(data, 2, mode, reg)
+	operand, offset, meta, err := decodeEA(data, inst.Address, 2, mode, reg)
 	if err != nil {
 		return err
 	}
@@ -66,7 +66,7 @@ func decodeJSR(data []byte, opcode uint16, inst *Instruction) error {
 func decodeJMP(data []byte, opcode uint16, inst *Instruction) error {
 	mode := uint8((opcode >> 3) & 0x7)
 	reg := uint8(opcode & 0x7)
-	operand, offset, meta, err := decodeEA(data, 2, mode, reg)
+	operand, offset, meta, err := decodeEA(data, inst.Address, 2, mode, reg)
 	if err != nil {
 		return err
 	}