+++ internal/decoders/special.go
@@ -9,7 +9,7 @@ func decodeLEA(data []byte, opcode uint16, inst *Instruction) error {
 	regX := uint8((opcode >> 9) & 0x7)
 	mode := uint8((opcode >> 3) & 0x7)
 	reg := uint8(opcode & 0x7)
-	operand, offset, meta, err := decodeEA(data, 2, mode, reg)
+	operand, offset, meta, err := decodeEA(data, inst.Address, 2, mode, reg)
 	if err != nil {
 		return err
 	}
@@ -20,7 +20,7 @@ func decodeLEA(data []byte, opcode uint16, inst *Instruction) error {
 func decodePEA(data []byte, opcode uint16, inst *Instruction) error {
 	mode := uint8((opcode >> 3) & 0x7)
 	reg := uint8(opcode & 0x7)
-	operand, offset, meta, err := decodeEA(data, 2, mode, reg)
+	operand, offset, meta, err := decodeEA(data, inst.Address, 2, mode, reg)
 	if err != nil {
 		return err
 	}
@@ -57,8 +57,22 @@ func decodeTRAPV(data []byte, opcode uint16, inst *Instruction) error {
 	return nil
 }
 
-func formatRegisterList(regListMask uint16) (string, []string) {
+func formatRegisterList(regListMask uint16, reverse bool) (string, []string) {
 	registers := []string{}
+	if reverse {
+		for i := 0; i < 8; i++ {
+			if regListMask&(1<<uint(15-i)) != 0 {
+				registers = append(registers, fmt.Sprintf("D%d", i))
+			}
+		}
+		for i := 0; i < 8; i++ {
+			if regListMask&(1<<uint(7-i)) != 0 {
+				registers = append(registers, fmt.Sprintf("A%d", i))
+			}
+		}
+		return formatRegisterRange(registers), registers
+	}
+
 	for i := 0; i < 8; i++ {
 		if regListMask&(1<<uint(i)) != 0 {
 			registers = append(registers, fmt.Sprintf("D%d", i))